pyproject.toml
[tool.poetry]
name = "work first ddd"
version = "0.1.0"
description = ""
authors = ["Birzhan Kairatov <birgan2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
django-filter = "^23.1"
pillow = "^9.5.0"
rest-framework-simplejwt = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
